% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildID.R
\name{buildID}
\alias{buildID}
\title{Build ID from columns of a data.frame.}
\usage{
buildID(idsDataFrame, sep = "/")
}
\arguments{
\item{idsDataFrame}{a data.frame.}

\item{sep}{the separator for the values in different columns.}
}
\value{
a vector containing the IDs.
}
\description{
This function builds ID from columns of a data frame: the columns of the data frame are
casted to character and then pasted.
}

