% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileSummary.R
\name{getFileSummary}
\alias{getFileSummary}
\title{Return summary on a STM data frame.}
\usage{
getFileSummary(dataFrame)
}
\arguments{
\item{dataFrame}{a data.frame with the supposed format.}
}
\value{
a data.frame with five columns:
LOTS: the number of different lots,
WAFERS: the number of different wafers,
LAYERS: the total number of different layers,
ROWS: the number of rows,
FAULTS: the total number of faulty chips.
}
\description{
This function return a data.frame containing some useful information about the data in a STM file.
}

