% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistoryFileSummary.R
\name{getHistoryFileSummary}
\alias{getHistoryFileSummary}
\title{Return summary on a STM lot processing history data frame.}
\usage{
getHistoryFileSummary(dataFrame)
}
\arguments{
\item{dataFrame}{a data.frame with the supposed format.}
}
\value{
a data.frame with five columns:
LOTS: the number of different lots,
OPERATIONS: the number of different operations,
EQUIPMENTS: the total number of different equipments,
START_DATE: the older date of an operation,
END_DATE: the most recent date of an operation
ROWS: the number of rows,
}
\description{
This function return a data.frame containing some useful 
information about the data in a STM lot processing history file.
}

