% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKDEData.R
\name{getKDEData}
\alias{getKDEData}
\title{Return a data.frame suitable for KDE algorithm.}
\usage{
getKDEData(dataFrame, waferIdentificativeColumns = c("LOT", "WAFER", "LAYER"),
  dieIdentificativeColumns = c("DIE_X", "DIE_Y"), xFaultColumn = "DIST_X",
  yFaultColumn = "DIST_Y", ...)
}
\arguments{
\item{dataFrame}{a data.frame with the supposed format.}

\item{waferIdentificativeColumns}{the columns that are to be used as identifiers of a wafer (ie. two wafers
are different if at least one of the value in these columns is different)}

\item{dieIdentificativeColumns}{the columns that are to be used as identifiers of a die (ie. two dies
are different if at least one of the value in these columns is different)}

\item{xFaultColumn}{the column contiaining the x coordinate of the fault.}

\item{yFaultColumn}{the column contiaining the y coordinate of the fault.}

\item{...}{additional parameters to be passed to buildID()}
}
\value{
a data.frame with three columns: ID,
FAULT_X (the x coordinate of the fault) and FAULT_Y (the y coordinate of the fault).
}
\description{
This function return a data.frame suitable for KDE algorithm.
It contains the identifier of each wafer and the (x,y) coordinates of each fault.
}

