% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLayerData.R
\name{getLayerData}
\alias{getLayerData}
\title{Selects the data of a single wafer by its lot, waferId and layer.}
\usage{
getLayerData(dataFrame, lot, wafer, layer)
}
\arguments{
\item{dataFrame}{a data.frame with the supposed format.}

\item{lot}{the lot of the wafer}

\item{wafer}{the ID of the wafer}

\item{layer}{the layer of the wafer}
}
\value{
a data.frame with the same layout as dataFrame.
}
\description{
This function selects the data of a single wafer by its lot, waferId and layer.
}

