% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLotData.R
\name{getLotData}
\alias{getLotData}
\title{Select the data of a single lot by its lot name.}
\usage{
getLotData(dataFrame, lot)
}
\arguments{
\item{dataFrame}{a data.frame with the supposed format.}

\item{lot}{the lot of the wafer}
}
\value{
a data.frame with the same layout as dataFrame.
}
\description{
This function selects the data of a single lot by its lot name.
}

