% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaferData.R
\name{getWaferData}
\alias{getWaferData}
\title{Select the data of a single wafer by its lot and waferId.}
\usage{
getWaferData(dataFrame, lot, wafer)
}
\arguments{
\item{dataFrame}{a data.frame with the supposed format.}

\item{lot}{the lot of the wafer}

\item{wafer}{the ID of the wafer}
}
\value{
a data.frame with the same layout as dataFrame.
}
\description{
This function selects the data of a single wafer by its lot and waferId.
}

