% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaferFeatures.R
\name{getWaferFeatures}
\alias{getWaferFeatures}
\title{Get dimensional features from data.frame}
\usage{
getWaferFeatures(dataFrame, dieWidthColumn = "DIE_WIDTH",
  dieHeightColumn = "DIE_HEIGHT", waferSizeColumn = "WAFER_SIZE")
}
\arguments{
\item{dataFrame}{a data.frame with the expected columns.}

\item{dieWidthColumn}{the name of the column in dataFrame containing die widths}

\item{dieHeightColumn}{the name of the column in dataFrame containing die heights}

\item{waferSizeColumn}{the name of the column in dataFrame containing wafer diameters}
}
\value{
a list of all the dimensional features of the wafer in the data.frame passed. In this list:
dieWidth: the width of a chip
dieHeight: the height of a chip
waferRay: the ray of the wafer
}
\description{
This function returns a list of all the dimensional features of the wafer in the data.frame passed.
}

