% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitID.R
\name{splitID}
\alias{splitID}
\title{Split string and put token in a data.frame}
\usage{
splitID(ids, sep = "/")
}
\arguments{
\item{ids}{: the string, or character vector to be splitted}

\item{sep}{: the separatorto be considered for splitting.}
}
\value{
a data.frame containing the tokens.
}
\description{
This function splits a string over a specific separator and puts 
the tokens in a data.frame
}

